requirements.txt
pandas
pre-commit
nbstripout
matplotlib
numpy
requests
requests-html
openpyxl
python-slugify
